coordinator/requirements.txt
# SafeSphere Coordinator — Python Dependencies
# Install with: pip install -r coordinator/requirements.txt

# API framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy>=2.0.30
aiosqlite>=0.20.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Push notifications (optional — mock mode if not installed)
# firebase-admin>=6.5.0

# WebSocket support (included in uvicorn[standard])
websockets>=12.0

# Already in main requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
